# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project_Alan API documentation example",
        "contact": {
            "email": "aldaberdiev99@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
...
        "/banners": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "Display all Banners",
                "operationId": "bannersAll",
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BannerShowRequest"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Banner not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Banners"
                ],
                "summary": "Store banner data",
                "description": "Store banner data",
                "operationId": "bannerStore",
                "parameters": [
                    {
                        "name": "description",
                        "in": "query",
                        "description": "The description of Banner",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "The title of Banner",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "photos",
                        "in": "query",
                        "description": "The photo of Banner",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass banner credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "price",
                                    "title"
                                ],
                                "properties": {
                                    "price": {
                                        "type": "integer",
                                        "example": "20"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "About actions"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/banners/{id}": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "Get banner by ID",
                "operationId": "bannersGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of banner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BannerShowRequest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Banner not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Banners"
                ],
                "summary": "Update banner by ID",
                "operationId": "bannersUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of banner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActionStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BannerShowRequest"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Banners"
                ],
                "summary": "Delete banner by ID",
                "operationId": "bannersDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of banner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Banner not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/banners/{id}/photos": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "Display all banner photos",
                "operationId": "photosProducts",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of banner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PhotoShowRequest"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Photos not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ActionStoreRequest": {
                "title": "Action post example",
                "description": "Some simple request createad as action",
                "properties": {
                    "price": {
                        "title": "price",
                        "description": "Price of action for storring",
                        "type": "integer",
                        "example": "55"
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title for storring",
                        "type": "string",
                        "example": "About action"
                    }
                },
                "type": "object"
            },
            "BannerShowRequest": {
                "title": "Banner get example",
                "description": "Some simple example",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Unique ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the banner",
                        "type": "string",
                        "example": "Description of banner"
                    },
                    "title": {
                        "title": "title",
                        "description": "Title of banner",
                        "type": "string",
                        "example": "About banner"
                    },
                    "created_at": {
                        "title": "Created_at",
                        "description": "Created time",
                        "type": "string",
                        "example": "2021-04-22T12:11:09.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated_at",
                        "description": "Updated time",
                        "type": "string",
                        "example": "2021-04-22T12:11:09.000000Z"
                    }
                },
                "type": "object"
            },
...
            "PhotoShowRequest": {
                "title": "Photo get example",
                "description": "Some simple example",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Unique ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "url": {
                        "title": "url",
                        "description": "Url of the Photo",
                        "type": "string",
                        "example": "Link of photo"
                    },
                    "created_at": {
                        "title": "Created_at",
                        "description": "Created time",
                        "type": "string",
                        "example": "2021-04-22T12:11:09.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated_at",
                        "description": "Updated time",
                        "type": "string",
                        "example": "2021-04-22T12:11:09.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "X-APP-ID": {
                "type": "apiKey",
                "name": "X-APP-ID",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}